pyproject.toml
[project]
name = "streaming-kafka"
version = "0.1.0"
description = ""
authors = [
    {name = "cmpbj",email = "magnocarlos3232@gmail.com"}
]
readme = "README.md"
requires-python = "3.9.20"
dependencies = [
    "confluent-kafka (>=2.8.0,<3.0.0)",
    "faker (>=36.1.1,<37.0.0)",
    "load-dotenv (>=0.1.0,<0.2.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "uuid (>=1.30,<2.0)",
    "azure-storage-blob (>=12.24.1,<13.0.0)",
    "azure-identity (>=1.20.0,<2.0.0)",
    "pyarrow (>=19.0.1,<20.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
